requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
pandas==2.2.0
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.2
scipy==1.12.0
joblib==1.3.2

# ── ML ────────────────────────────────────────────────────────────────────────
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.0
shap==0.44.1
optuna==3.5.0

# ── MLOps ─────────────────────────────────────────────────────────────────────
mlflow==2.11.0

# ── Data ─────────────────────────────────────────────────────────────────────
faker==22.5.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.32.0
plotly==5.20.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.0.2
